refactor(crm): consolidate v2 api files and merge board api
Drop the redundant Crm prefix from the v2 api filenames to mirror the v1 module
(CrmCompanyApi -> CompanyApi, CrmDealApi -> DealApi), merge the two board files
(BoardApi's init-data hook + CrmBoardApi's grouped/reorder/move) into a single
BoardApi.ts, and rename the standalone LookupApi -> ContactApi since its two
typeahead calls hit the /contact/* routes (v1 keeps them in ContactApi too).
Import sites updated across the v2 hooks and components.
