+++ frontend/src/community/crm/v2/components/molecules/DeleteDealModalV2/DeleteDealModalV2.tsx
@@ -5,7 +5,7 @@ import { ToastType } from "~community/common/enums/ComponentEnums";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { useToast } from "~community/common/providers/ToastProvider";
 import CrmDeleteModalContent from "~community/crm/components/molecules/CrmDeleteModalContent/CrmDeleteModalContent";
-import { useDeleteDeal } from "~community/crm/v2/api/CrmDealApi";
+import { useDeleteDeal } from "~community/crm/v2/api/DealApi";
 import { useCrmStoreV2 } from "~community/crm/v2/store/store";
 import { removeDeal } from "~community/crm/v2/utils/dealIngest";
