+++ frontend/src/community/crm/v2/components/organisms/AddDealSidePanelV2/AddDealSidePanelV2.tsx
@@ -11,11 +11,11 @@ import {
   DEFAULT_LOOKUP_PAGE_SIZE,
   SEARCH_DEBOUNCE_DELAY
 } from "~community/crm/constants/commonConstants";
-import { useGetContactLookupV2 } from "~community/crm/v2/api/CrmLookupApi";
+import { useGetContactLookupV2 } from "~community/crm/v2/api/ContactApi";
 import {
   useCheckDealNameExists,
   useCreateDeal
-} from "~community/crm/v2/api/CrmDealApi";
+} from "~community/crm/v2/api/DealApi";
 import { CrmPriorityEnum } from "~community/crm/v2/enums/common";
 import { useCrmStoreV2 } from "~community/crm/v2/store/store";
 import { CrmDealEntity } from "~community/crm/v2/types/CrmCommonTypes";
