+++ frontend/src/community/crm/v2/components/organisms/DealSidePanelV2/DealTitleSection.tsx
@@ -10,7 +10,7 @@ import useDebounce from "~community/common/hooks/useDebounce";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { SEARCH_DEBOUNCE_DELAY } from "~community/crm/constants/commonConstants";
 import useInlineEditForm from "~community/crm/hooks/useInlineEditForm";
-import { useCheckDealNameExists } from "~community/crm/v2/api/CrmDealApi";
+import { useCheckDealNameExists } from "~community/crm/v2/api/DealApi";
 import { validateDealName } from "~community/crm/v2/utils/dealValidations";
 
 interface DealTitleSectionProps {
